requirements.txt
langchain-groq>=0.0.3
langchain-community>=0.0.13
python-dotenv>=1.0.0
docx2txt>=0.8
langchain>=0.1.0